+++ test/models/shipit/deploy_spec/file_system_test.rb
@@ -7,7 +7,8 @@ module Shipit
     class FileSystemTest < ActiveSupport::TestCase
       test 'deploy.pre calls "exit 1" if there is a bare shipit file and Shipit is configured to ignore' do
         Shipit.expects(:respect_bare_shipit_file?).returns(false).at_least_once
-        deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, 'env')
+        stack = shipit_stacks(:shipit)
+        deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, stack)
         deploy_spec.expects(:config_file_path).returns(Pathname.new(Dir.tmpdir) + '/shipit.yml').at_least_once
         deploy_spec.expects(:read_config).never
         pre_commands = deploy_spec.send(:config, 'deploy', 'pre')
@@ -18,7 +19,8 @@ module Shipit
 
       test 'deploy.pre does not call "exit 1" if Shipit is not configured to do so' do
         Shipit.expects(:respect_bare_shipit_file?).returns(true).at_least_once
-        deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, 'env')
+        stack = shipit_stacks(:shipit)
+        deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, stack)
         deploy_spec.expects(:config_file_path).returns(Pathname.new(Dir.tmpdir) + '/shipit.yml').at_least_once
         deploy_spec.expects(:read_config).returns(SafeYAML.load(deploy_spec_yaml))
         pre_commands = deploy_spec.send(:config, 'deploy', 'pre')
@@ -29,7 +31,8 @@ module Shipit
       test 'Shipit.respect_bare_shipit_file? has no effect if the file is not a bare file' do
         [true, false].each do |obey_val|
           Shipit.expects(:respect_bare_shipit_file?).returns(obey_val).at_least_once
-          deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, 'env')
+          stack = shipit_stacks(:shipit)
+          deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, stack)
           deploy_spec.expects(:config_file_path).returns(Pathname.new(Dir.tmpdir) + '/shipit.env.yml').at_least_once
           deploy_spec.expects(:read_config).returns(SafeYAML.load(deploy_spec_yaml))
           pre_commands = deploy_spec.send(:config, 'deploy', 'pre')
@@ -40,7 +43,8 @@ module Shipit
 
       test '#load_config does not error if the file is empty' do
         Shipit.expects(:respect_bare_shipit_file?).returns(true).at_least_once
-        deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, 'env')
+        stack = shipit_stacks(:shipit)
+        deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, stack)
         deploy_spec.expects(:config_file_path).returns(Pathname.new(Dir.tmpdir) + '/shipit.env.yml').at_least_once
         deploy_spec.expects(:read_config).at_least_once.returns(false)
         loaded_config = deploy_spec.send(:cacheable_config)
@@ -49,7 +53,8 @@ module Shipit
 
       test '#load_config does not error if there is no "deploy" key' do
         Shipit.expects(:respect_bare_shipit_file?).returns(false).at_least_once
-        deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, 'env')
+        stack = shipit_stacks(:shipit)
+        deploy_spec = Shipit::DeploySpec::FileSystem.new(Dir.tmpdir, stack)
         deploy_spec.expects(:config_file_path).returns(Pathname.new(Dir.tmpdir) + '/shipit.yml').at_least_once
         deploy_spec.expects(:read_config).never
         loaded_config = deploy_spec.send(:load_config)